requirements.txt
aiogram==3.12.0
asyncpg==0.31.0
python-dotenv==1.2.1
sqlalchemy==2.0.46
pydantic_settings==2.12.0